rSON: Restructured the code base to make nagivating it easier and modernise the naming style
