+++ include/meson.build
@@ -1,0 +1,12 @@
+# SPDX-License-Identifier: LGPL-3.0-or-later
+# SPDX-FileCopyrightText: 2018-2023 Rachel Mant <git@dragonmux.network>
+# SPDX-FileContributor: Written by Rachel Mant <git@dragonmux.network>
+
+install_headers(
+	'rSON.hxx'
+)
+
+rSONIncludeDir = include_directories('.')
+rSONIncludePath = meson.current_source_dir()
+
+subdir('socket')
