+++ include/socket/meson.build
@@ -1,0 +1,9 @@
+# SPDX-License-Identifier: LGPL-3.0-or-later
+# SPDX-FileCopyrightText: 2018-2023 Rachel Mant <git@dragonmux.network>
+# SPDX-FileContributor: Written by Rachel Mant <git@dragonmux.network>
+
+install_headers(
+	'rSONSocket.hxx'
+)
+
+rSONSocketIncludeDir = include_directories('.')
