--- rSON_socket.pc.in
@@ -1,30 +1,0 @@
-# This file is part of rSON
-# Copyright © 2017 Rachel Mant (dx-mon@users.sourceforge.net)
-#
-# rSON is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Lesser General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# rSON is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Lesser General Public License for more details.
-#
-# You should have received a copy of the GNU Lesser General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-# rSON pkg-config file
-
-prefix=@PREFIX@
-exec_prefix=${prefix}
-libdir=@LIBDIR@
-includedir=${prefix}/include
-
-Name: rSON Socket
-Description: Sockets handling for rSON, a simple but effective JSON parser for C++
-Version: 0.2.2
-Requires:
-Conflicts:
-Libs: -L${libdir} -lrSON.socket -lrSON
-Cflags: -I${includedir}
